requirements.txt
streamlit
opencv-python-headless
numpy
pandas
plotly
Pillow